# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Codeplane API
  version: 1.0.0
  description: API for Codeplane - a jj-native development platform
servers:
  - url: https://api.codeplane.app
paths:
...
  /repos/{owner}/{repo}/agent/sessions:
    get:
      summary: GET /repos/{owner}/{repo}/agent/sessions
      description: List agent sessions for a repository.
      operationId: get_repos__owner___repo__agent_sessions
      tags:
        - Agent
      parameters:
        - name: owner
          in: path
          required: true
          schema:
            type: string
        - name: repo
          in: path
          required: true
          schema:
            type: string
        - name: cursor
          in: query
          description: Opaque cursor from a previous response's next_cursor field
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 30
            maximum: 100
      responses:
        '200':
          description: List of agent sessions
          headers:
            Link:
              description: RFC 8288 pagination links with cursor-based URLs.
              schema:
                type: string
          content:
            application/json:
              schema:
                type: object
                required:
                  - items
                  - next_cursor
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/AgentSession'
                  next_cursor:
                    type: string
                    nullable: true
                    description: Opaque cursor for fetching the next page. Null when there are no more results.
    post:
      summary: POST /repos/{owner}/{repo}/agent/sessions
      description: Create a new agent session for the repository.
      operationId: post_repos__owner___repo__agent_sessions
      tags:
        - Agent
      parameters:
        - name: owner
          in: path
          required: true
          schema:
            type: string
        - name: repo
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                  description: Optional session title.
      responses:
        '201':
          description: Agent session created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgentSession'
  /repos/{owner}/{repo}/agent/sessions/{id}:
    get:
      summary: GET /repos/{owner}/{repo}/agent/sessions/{id}
      description: Get a single agent session.
      operationId: get_repos__owner___repo__agent_sessions__id_
      tags:
        - Agent
      parameters:
        - name: owner
          in: path
          required: true
          schema:
            type: string
        - name: repo
          in: path
          required: true
          schema:
            type: string
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Agent session details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgentSession'
  /repos/{owner}/{repo}/agent/sessions/{id}/messages:
    get:
      summary: GET /repos/{owner}/{repo}/agent/sessions/{id}/messages
      description: List messages for an agent session in sequence order.
      operationId: get_repos__owner___repo__agent_sessions__id__messages
      tags:
        - Agent
      parameters:
        - name: owner
          in: path
          required: true
          schema:
            type: string
        - name: repo
          in: path
          required: true
          schema:
            type: string
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: cursor
          in: query
          description: Opaque cursor from a previous response's next_cursor field
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 30
            maximum: 100
      responses:
        '200':
          description: List of agent messages
          headers:
            Link:
              description: RFC 8288 pagination links with cursor-based URLs.
              schema:
                type: string
          content:
            application/json:
              schema:
                type: object
                required:
                  - items
                  - next_cursor
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/AgentMessage'
                  next_cursor:
                    type: string
                    nullable: true
                    description: Opaque cursor for fetching the next page. Null when there are no more results.
    post:
      summary: POST /repos/{owner}/{repo}/agent/sessions/{id}/messages
      description: Append a message to an agent session. Posting a `user` message dispatches a cloud VM-backed agent run.
      operationId: post_repos__owner___repo__agent_sessions__id__messages
      tags:
        - Agent
      parameters:
        - name: owner
          in: path
          required: true
          schema:
            type: string
        - name: repo
          in: path
          required: true
          schema:
            type: string
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - role
                - parts
              properties:
                role:
                  type: string
                  enum: [user, assistant, system, tool]
                parts:
                  type: array
                  items:
                    type: object
                    required:
                      - type
                      - content
                    properties:
                      type:
                        type: string
                        enum: [text, tool_call, tool_result]
                      content:
                        description: Message part payload. Text parts accept either a string or an object.
      responses:
        '201':
          description: Agent message created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgentMessage'
  /repos/{owner}/{repo}/agent/sessions/{id}/stream:
    get:
      summary: GET /repos/{owner}/{repo}/agent/sessions/{id}/stream
      description: SSE endpoint for agent session output and terminal done events.
      operationId: get_repos__owner___repo__agent_sessions__id__stream
      tags:
        - Agent
      parameters:
        - name: owner
          in: path
          required: true
          schema:
            type: string
        - name: repo
          in: path
          required: true
          schema:
            type: string
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: SSE event stream
          content:
            text/event-stream:
              schema:
                type: string
...
components:
  schemas:
    AgentSession:
      type: object
      properties:
        id:
          type: string
          format: uuid
        repository_id:
          type: integer
          format: int64
        user_id:
          type: integer
          format: int64
        title:
          type: string
        status:
          type: string
          enum: [active, completed, failed]
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    AgentMessage:
      type: object
      properties:
        id:
          type: integer
          format: int64
        session_id:
          type: string
          format: uuid
        role:
          type: string
          enum: [user, assistant, system, tool]
        sequence:
          type: integer
          format: int64
        parts:
          type: array
          items:
            $ref: '#/components/schemas/AgentMessagePart'
        created_at:
          type: string
          format: date-time
    AgentMessagePart:
      type: object
      properties:
        part_index:
          type: integer
          format: int64
        type:
          type: string
          enum: [text, tool_call, tool_result, done]
        content:
          description: Message part payload.
...